handle error - member controllers

+++ src/controllers/memberControllers.ts
@@ -1,32 +1,52 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 import { UserModel } from "../models/userModel";
 import { CommunityModel } from "../models/communityModel";
 import { RoleModel } from "../models/roleModel";
 import { MemberModel } from "../models/memberModel";
 import { Snowflake } from "@theinternetfolks/snowflake";
-import { canRemoveIn } from "../utils/canIn";
+import { canAddIn, canRemoveIn } from "../utils/canIn";
+import { createErrMessage } from "../utils/errors";
 
-export const addMember = async (req: Request, res: Response) => {
+export const addMember = async (req: Request, res: Response, next: NextFunction) => {
   const { community, user, role } = req.body;
   const created_at = new Date().toISOString();
   const id = Snowflake.generate();
+
+  const errorArray = [];
 
   try {
     // validate
     const communityExists = await CommunityModel.findOne({ id: community });
     if (!communityExists) {
-      return res.status(400).json({ message: "community does not exist" });
+      const err = createErrMessage({ code: "RESOURCE_NOT_FOUND", param: "community" });
+      errorArray.push(err);
     }
 
     const userExits = await UserModel.findOne({ id: user });
     if (!userExits) {
-      return res.status(400).json({ message: "user does not exist" });
+      const err = createErrMessage({ code: "RESOURCE_NOT_FOUND", param: "user" });
+      errorArray.push(err);
     }
 
     const roleExists = await RoleModel.findOne({ id: role });
     if (!roleExists) {
-      return res.status(400).json({ message: "role does not exist" });
+      const err = createErrMessage({ code: "RESOURCE_NOT_FOUND", param: "role" });
+      errorArray.push(err);
     }
+
+    const userAlreadyAMember = await MemberModel.findOne({ community: community, user: user });
+    if (userAlreadyAMember) {
+      const err = createErrMessage({ code: "RESOURCE_EXISTS" });
+      errorArray.push(err);
+    }
+
+    const canAddInCommunities = await canAddIn(req);
+    if (!canAddInCommunities.includes(community)) {
+      const err = createErrMessage({ code: "NOT_ALLOWED_ACCESS" });
+      errorArray.push(err);
+    }
+
+    if (errorArray.length > 0) return next(errorArray);
 
     // Add member
     const member = await MemberModel.create({
@@ -48,12 +68,26 @@ export const addMember = async (req: Request, res: Response) => {
   }
 };
 
-export const removeMember = async (req: Request, res: Response) => {
+export const removeMember = async (req: Request, res: Response, next: NextFunction) => {
   if (!req.params.id) return res.status(400).json({ message: "id is required" });
   const userIdToDelete = req.params.id;
 
+  const errorArray = [];
   try {
+    const userExits = await UserModel.findOne({ id: userIdToDelete });
+    if (!userExits) {
+      const err = { code: "RESOURCE_NOT_FOUND" };
+      errorArray.push(err);
+    }
+
     const communitiesToRemoveFrom = await canRemoveIn(req);
+
+    // +++++++
+    // check if
+
+    // +++++++
+
+    if (errorArray.length > 0) return next(errorArray);
 
     for (const community of communitiesToRemoveFrom) {
       await MemberModel.deleteOne({ community: community, user: userIdToDelete });
